refactor(VehiclePlate): added constants & model, refactored file name & helper methods
