+++ lib/src/constants/vehicle_plate/constants.dart
@@ -1,3 +1,5 @@
+const nonDigitRegExp = r'\D';
+
 const plateDataset = [
   {
     10: 'تهران',
