+++ lib/src/core/vehicle_plate/model.dart
@@ -1,0 +1,20 @@
+class NormalizedPlate {
+  String numbers;
+  String? char;
+
+  NormalizedPlate(this.numbers, [this.char]);
+}
+
+class PlateInfo {
+  String template;
+  String? province;
+  String type;
+  String? category;
+
+  PlateInfo({
+    required this.template,
+    this.province,
+    required this.type,
+    this.category,
+  });
+}
